refatoração para utilizar repository de legislação
