+++ app/Http/Controllers/Site/HomeController.php
@@ -2,14 +2,11 @@
 
 namespace App\Http\Controllers\Site;
 
-use App\Models\Site\LegislacaoTipo;
 use App\Http\Controllers\Controller;
-use App\Models\Site\Legislacao;
-use Illuminate\Http\Request;
 use App\Models\Site\Slider;
 use App\Models\Site\Noticia;
 use App\Models\Site\Parlamentar;
-use App\Models\Site\LegislacaoSituacao;
+use App\Repositories\LegislacaoRepository;
 use Illuminate\Support\Facades\DB;
 
 class HomeController extends Controller
@@ -19,7 +16,7 @@ class HomeController extends Controller
         return redirect()->route('home');
     }
 
-    public function indexSite()
+    public function indexSite(LegislacaoRepository $legislacaoRepository)
     {
         $contslider = 0;
         $slider = Slider::all() ->where('exibir', 1);
@@ -55,65 +52,27 @@ class HomeController extends Controller
                                                     ->join('sitefox_vereador_legislatura AS svl', 'sv.id', '=', 'svl.id_vereador')
                                                     ->join('sitefox_legislatura AS sl', 'sl.id', '=', 'svl.id_legislatura')
                                                     ->get();
+
+        $especiesNormativas = $legislacaoRepository->getLegislacoesPorTipo(1);
+        $legislacoesTema = $legislacaoRepository->getLegislacoesPorTipo(2);
+                                                    
+        $legislacao['especiesNormativas'] = $especiesNormativas;
+        $legislacao['temas'] = $legislacoesTema;
+        $legislacao['autores'] = $legislacaoRepository->getLegislacoesPorVereador();
+        $legislacao['situacoes'] = $legislacaoRepository->getLegislacoesPorSituacao();
         
-        $legislacao['especiesNormativas'] = $this->getLegislacoesPorTipo(1);
-        $legislacao['temas'] = $this->getLegislacoesPorTipo(2);
-        $legislacao['autores'] = $this->getLegislacoesPorVereador();
-        $legislacao['situacoes'] = $this->getLegislacoesPorSituacao();
-                                            
-        $especiesNormativas = $this->getLegislacoesPorTipo(1);
-        $legislacoesTema = $this->getLegislacoesPorTipo(2);
-
-        return view('site.home', ['contslider' => $contslider,
-                                  'slider' => $slider,
-                                  'destaque' => $destaque,
-                                  'noticia' =>$noticia,
-                                  'videos' => $videos,
-                                  'videos_d' => $videos_d,
-                                  'parlamentares' => $parlamentares,
-                                  'presidente' => $presidente,
-                                  'legislacao' => $legislacao,
-                                  'especiesNormativas'=> $especiesNormativas,
-                                  'legislacoesTema' => $legislacoesTema
-                                    ]);
-    }
-
-    protected function getLegislacoes(int $tipo)
-    {
-        return Legislacao::selectRaw('sitefox_legislacao_tipo.nome, sitefox_legislacao_tipo.id, count(sitefox_legislacao_tipo.id) as count')
-        ->join('sitefox_legislacao_tipo', 'sitefox_legislacao.id_tipo', 'sitefox_legislacao_tipo.id')
-        ->groupBy('sitefox_legislacao_tipo.nome')
-        ->groupBy('sitefox_legislacao_tipo.id')
-        ->where('sitefox_legislacao_tipo.tipo', $tipo)
-        ->get();
-    }
-
-    protected function getLegislacoesPorVereador()
-    {
-        return Legislacao::selectRaw('sitefox_vereador.id, sitefox_vereador.nome, count(sitefox_vereador.id) as count')
-        ->join('sitefox_vereador', 'sitefox_legislacao.id_vereador', 'sitefox_vereador.id')
-        ->groupBy('sitefox_vereador.id')
-        ->groupBy('sitefox_vereador.nome')
-        ->get();
-    }
-
-    protected function getLegislacoesPorSituacao()
-    {
-        return Legislacao::selectRaw('sitefox_legislacao_situacao.id, sitefox_legislacao_situacao.nome, count(sitefox_legislacao_situacao.id) as count')
-        ->join('sitefox_legislacao_situacao', 'sitefox_legislacao.id_situacao', 'sitefox_legislacao_situacao.id')
-        ->groupBy('sitefox_legislacao_situacao.id')
-        ->groupBy('sitefox_legislacao_situacao.nome')
-        ->get();
-    }
-
-    protected function getLegislacoesPorTipo(int $tipo)
-    {
-        return Legislacao::selectRaw('sitefox_legislacao_tipo.nome, sitefox_legislacao_tipo.id, count(sitefox_legislacao_tipo.id) as count')
-        ->join('sitefox_legislacao_tipo', 'sitefox_legislacao.id_tipo', 'sitefox_legislacao_tipo.id')
-        ->groupBy('sitefox_legislacao_tipo.nome')
-        ->groupBy('sitefox_legislacao_tipo.id')
-        ->where('sitefox_legislacao.ativo', 1)
-        ->where('sitefox_legislacao_tipo.tipo', $tipo)
-        ->get();
+        return view('site.home', [
+            'contslider' => $contslider,
+            'slider' => $slider,
+            'destaque' => $destaque,
+            'noticia' =>$noticia,
+            'videos' => $videos,
+            'videos_d' => $videos_d,
+            'parlamentares' => $parlamentares,
+            'presidente' => $presidente,
+            'legislacao' => $legislacao,
+            'especiesNormativas'=> $especiesNormativas,
+            'legislacoesTema' => $legislacoesTema
+        ]);
     }
 }
